+++ src/assets/sass/plugins/sdk-sc-explorer/_card-item.scss
@@ -1,0 +1,72 @@
+@import './variables';
+
+/* ----------------- Card Item ----------------- */
+
+.#{$package-name} {
+  &-card-item {
+    row-gap: 0.5rem;
+    column-gap: 1rem;
+    padding: 0.25rem 0.5rem;
+    flex-basis: 30%;
+    flex-grow: 0;
+    overflow: hidden;
+
+    &:last-of-type {
+      flex-grow: 1;
+    }
+    &-container {
+      display: flex;
+      flex-wrap: wrap;
+      row-gap: 1rem;
+      @media (max-width: 767px) {
+        flex-direction: column;
+      }
+    }
+    &-icon {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+
+      background-color: transparent;
+      box-shadow: 0 0 0.25rem 0 rgba(35, 247, 221, 0.3);
+
+      border-radius: 0.5rem;
+      min-width: 3.2rem;
+      min-height: 3.2rem;
+
+      svg {
+        width: 1.15rem !important;
+        height: 1.15rem !important;
+        fill: var(--primary);
+        color: var(--primary);
+      }
+    }
+    &-content {
+      min-width: 0;
+      width: 100%;
+    }
+    &-title {
+      color: var(--neutral-400);
+    }
+    &-value {
+      .dapp-trim {
+        display: flex;
+        max-width: 100%;
+        overflow: hidden;
+        position: relative;
+        white-space: nowrap;
+      }
+      .dapp-copy-button {
+        margin-left: 0.5rem;
+        svg[data-icon='copy'] {
+          color: var(--neutral-500);
+        }
+      }
+    }
+  }
+  &-collapsible-card {
+    &.#{$package-name}-card > .#{$package-name}-card-header {
+      cursor: pointer;
+    }
+  }
+}
