+++ src/assets/sass/plugins/sdk-sc-explorer/_code-block.scss
@@ -1,0 +1,45 @@
+@import './variables';
+
+/* ----------------- Code Block ----------------- */
+
+.#{$package-name} {
+  &-code-block {
+    position: relative;
+    & > pre {
+      min-height: 2.813rem;
+      background: var(--table-header) !important;
+      border-radius: 0.5rem;
+      margin-bottom: 0;
+      overflow-x: auto;
+      & > code {
+        display: block;
+        word-break: break-all;
+      }
+    }
+  }
+  &-button-holder {
+    position: absolute;
+    right: 0.5rem;
+    top: 0.5rem;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+  }
+  &-copy-button,
+  &-link-button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 1.813rem;
+    width: 1.813rem;
+    border-radius: 0.5rem;
+    margin-left: 0;
+    background-color: var(--neutral-900);
+    &:hover {
+      background-color: var(--black);
+    }
+    svg {
+      height: 0.75rem;
+    }
+  }
+}
