+++ src/assets/sass/plugins/sdk-sc-explorer/_react-bootstrap.scss
@@ -1,0 +1,41 @@
+@import './variables';
+
+/* ----------------- React Bootstrap Helpers ----------------- */
+
+.#{$package-name} {
+  // Collapse
+  .fade {
+    transition: opacity 0.15s linear;
+  }
+  @media (prefers-reduced-motion: reduce) {
+    .fade {
+      transition: none;
+    }
+  }
+  .fade:not(.show) {
+    opacity: 0;
+  }
+  .collapse:not(.show) {
+    display: none;
+  }
+  .collapsing {
+    height: 0;
+    overflow: hidden;
+    transition: height 0.35s ease;
+  }
+  @media (prefers-reduced-motion: reduce) {
+    .collapsing {
+      transition: none;
+    }
+  }
+  .collapsing.collapse-horizontal {
+    width: 0;
+    height: auto;
+    transition: width 0.35s ease;
+  }
+  @media (prefers-reduced-motion: reduce) {
+    .collapsing.collapse-horizontal {
+      transition: none;
+    }
+  }
+}
