+++ src/assets/sass/plugins/sdk-sc-explorer/_sections.scss
@@ -1,0 +1,120 @@
+@import './variables';
+
+/* ----------------- Sections ----------------- */
+
+.#{$package-name} {
+  // Build Info
+  &-contract-build {
+    display: flex;
+    flex-direction: column;
+    .#{$package-name}-card-body {
+      --card-spacer-x: 0;
+      --card-spacer-y: 0;
+    }
+  }
+
+  // Endpoints
+  &-endpoint-output {
+    .#{$package-name}-panel-content {
+      @media (min-width: 768px) {
+        padding: 0.5rem 0 0 0.5rem;
+      }
+    }
+    &-results {
+      pre {
+        border-radius: var(--border-radius);
+        margin-top: 0.5rem;
+        margin-bottom: 0;
+      }
+      @media (min-width: 768px) {
+        padding-right: 0.5rem;
+        padding-bottom: 0.5rem;
+      }
+    }
+  }
+  &-endpoint-general-error {
+    @media (min-width: 768px) {
+      margin-top: 0.5rem;
+      margin-left: calc(10% + 0.5rem);
+    }
+  }
+  &-endpoint-action-wrapper {
+    display: flex;
+    flex-wrap: wrap;
+    column-gap: 1rem;
+    row-gap: 1rem;
+    margin-top: 1rem;
+    @media (min-width: 768px) {
+      margin-top: 0.5rem;
+      margin-left: calc(10% + 0.5rem);
+    }
+  }
+
+  // Constructor
+  &-contract-constructor {
+    .#{$package-name}-card-body {
+      padding-top: 0.875rem;
+      @media (max-width: 767px) {
+        --card-spacer-y: 0.5rem;
+      }
+    }
+    .#{$package-name}-panel-content {
+      padding-top: 0.625rem;
+    }
+  }
+  &-about-panel {
+    margin-top: 1rem;
+  }
+  &-about-panel-content {
+    padding: 0.625rem 1rem;
+    border-radius: var(--border-radius);
+    background-color: var(--neutral-900);
+    p {
+      margin-bottom: 0;
+    }
+  }
+
+  // Load ABI
+  &-deploy-upgrade-file-form,
+  &-contract-load-abi-form {
+    &-fields {
+      margin-top: 1.5rem;
+      margin-bottom: 1.5rem;
+      .#{$package-name}-input-group-append {
+        padding-left: 1rem;
+        padding-right: 1rem;
+        color: var(--primary);
+      }
+    }
+    &-code-container {
+      margin-top: 1.5rem;
+      .#{$package-name}-code {
+        max-height: 20rem;
+        overflow-y: auto;
+      }
+    }
+  }
+
+  // Deploy/Upgrade Contract
+  &-deploy-upgrade-file-form {
+    .#{$package-name}-form-action-wrapper {
+      margin-top: 1.5rem;
+      margin-bottom: 1.5rem;
+    }
+    .#{$package-name}-dropzone {
+      margin-bottom: 1.5rem;
+    }
+  }
+
+  &-interaction-modal-form {
+    &-code-metadata {
+      display: flex;
+      align-items: center;
+      flex-wrap: wrap;
+      column-gap: 1rem;
+      row-gap: 0.5rem;
+      margin-top: 1.5rem;
+      margin-bottom: 1.5rem;
+    }
+  }
+}
