+++ src/assets/sass/plugins/sdk-sc-explorer/_variables.scss
@@ -1,0 +1,8 @@
+$package-name: 'mx-sdk-sc';
+$font-size-base: 1rem;
+$headings-font-family: 'Roobert Medium', 'Inter', Roboto, 'Helvetica Neue',
+  'Noto Sans', 'Liberation Sans', sans-serif;
+$font-family-monospace: SFMono-Regular, Menlo, Monaco, Consolas,
+  'Liberation Mono', 'Courier New', monospace;
+$spacer-val: 2rem;
+$table-header: #1d1d1d;
