+++ src/assets/sass/theme.scss
@@ -164,4 +164,4 @@ body {
   }
 }
 
-@import './plugins/sdk-sc-explorer.scss';
+@import './plugins/sdk-sc-explorer/sdk-sc-explorer.scss';
