+++ src/pages/SmartContractInteraction/index.ts
@@ -1,0 +1,1 @@
+export * from './SmartContractInteraction';
