+++ src/pages/SmartContractInteraction/styles.module.scss
